requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
Telethon==1.33.1
python-dotenv==1.0.1
feedparser==6.0.11
flask-cors==4.0.0
deep-translator==1.11.4
